Modify whitespace trimming to be less hungry
The liquid spec is a little vague about quite how greedy whitespace
trimming should be. Currently handling here will eat any available
whitespace between tags including all newline characters
e.g.
```
(

{%- sometag -%}

)
```
will render as
```
()
```
This is ok for web markup where the whitespace is not important, but for
other document types this greedy consumption is problematic.

Ideally the ws trimming would be less hungry and only trim any whitespace
on the line containing the tag up to and including it's newline, essentially
removing any trace of the tag but leaving other lines intact rendering the above as
```
(

)
```

The change here detects whitespace as tokens as they are parsed and
discards any as indicated by the trimming directives up to an including
the first newline.