Fix slice filter on strings longer than 1000 runes
The previous implementation compiled a regex on the fly, using a
repetition count that could go above the supported range, causing it to
panic.

This implementation should work on any string size, offset and length.

I added edge case tests when using out-of-bounds values. I do not know
what Shopify implementation would return, and the documentation does not
specify any behavior, so I came up with a sensible behavior.