Add support for strict variables
The Ruby implementation has support for erroring when a template has an
undefined variable. This is implemented by passing an option to the
render() method.

As this version doesn't expose render.Config in the engine, a
StrictVariables() method is provided on the engine to enable it. This
differs from the Ruby version in that it's on for the entire engine,
rather than enabled for each call to Render(). This is more similar to
the Ruby version's render!() method as it immediately errors, rather
than storing a stack of errors which can be accessed later.

Refs. #8